+++ src/sdk/CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added link unfurling: throw any SharePoint (sharing) link in and you will get back detailed information about the SharePoint resource, including thumbnails #847 [jansenbe - Bert Jansen]
 - Added sharing capabilities (add) for list items #879 [jansenbe - Bert Jansen]
 - Added support for adding, publishing and unpublishing content types and document sets from the tenant's content type hub #881 [MathijsVerbeeck - Mathijs Verbeeck]
+- Added support for consuming content types from the content type hub [jansenbe - Bert Jansen]
 
 ### Changed